Change .data pseudo-instruction to allow easily repeated values.
If a .data instruction has too few values inside the brackets, the values in the brackets are cycled through as many times as necessary to fill the desired number of cells.  If a .data instruction has too many values inside the brackets, some of the rightmost values inside the brackets are not stored in cells.