Add Matchers and Visitor pattern to Machine
This commit adds `Matcher` values for testing `Machine` components. These
Matchers are reusable and allow for more distinct testing of components.
The "overall" methods do not test that the ID matches as often we are
building "expected" objects that match everything _except_ the
matcher.

This only includes `Module`s, and is not fully complete. There are still
`Microinstruction`s missing and need to be tested more.

Additionally, this adds a `MachineBuilder` for repeatably specifying
`Machine` values in reproducable ways. There is a `VerificationBuilder` in
the works that will also verify that the machine is built correctly
according to the specifications -- the `MachineBuilder.Default` version
does not verify contents and assumes they are correct.

In addition, there are now "visitor" pattern-style traversals available to
the `Machine`. These are not tested yet and are a bit of a work in
progress.

* Changed all components that can be `null` to return `Optional` and
adjust methods calls where required
* Fixed some properties to make them named better
* Completed implementation of all Microinstruction matchers